Moves reference PWV values into the sn_magnitudes module
